+++ src/main/java/racingcargame/Application.java
@@ -1,8 +1,8 @@
 package racingcargame;
 
 import racingcargame.controller.GameController;
-import racingcargame.model.movementgenerator.MoveGenerator;
-import racingcargame.model.numbergenerator.RandomNumberGenerator;
+import racingcargame.model.carmovevalue.MoveValueGeneratorByRandomNumber;
+import racingcargame.model.carmovevalue.randomnumbergenerator.RandomNumberGenerator;
 import racingcargame.view.InputView;
 import racingcargame.view.OutputView;
 
@@ -10,7 +10,7 @@ public class Application {
 
     public static void main(String[] args) {
         GameController gameController = new GameController(new InputView(), new OutputView(),
-                new MoveGenerator(new RandomNumberGenerator()));
+                new MoveValueGeneratorByRandomNumber(new RandomNumberGenerator()));
         gameController.runGame();
     }
 }
