+++ src/main/java/racingcargame/model/carmovevalue/MoveValueGenerator.java
@@ -1,0 +1,5 @@
+package racingcargame.model.carmovevalue;
+
+public interface MoveValueGenerator {
+    int getMoveValue();
+}
