+++ src/main/java/racingcargame/model/carmovevalue/MoveValueGeneratorByRandomNumber.java
@@ -1,15 +1,14 @@
-package racingcargame.model.movementgenerator;
+package racingcargame.model.carmovevalue;
 
-import racingcargame.model.movementgenerator.Move;
-import racingcargame.model.numbergenerator.NumberGenerator;
+import racingcargame.model.carmovevalue.randomnumbergenerator.NumberGenerator;
 
-public class MoveGenerator implements Move {
+public class MoveValueGeneratorByRandomNumber implements MoveValueGenerator {
     private static final int MAX_BOUNDARY_NUMBER = 10;
     private static final int NUMBER_MAKE_MOVE_POSSIBLE = 4;
 
     private final NumberGenerator numberGenerator;
 
-    public MoveGenerator(final NumberGenerator numberGenerator) {
+    public MoveValueGeneratorByRandomNumber(final NumberGenerator numberGenerator) {
         this.numberGenerator = numberGenerator;
     }
