+++ src/main/java/racingcargame/model/carmovevalue/randomnumbergenerator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcargame.model.numbergenerator;
+package racingcargame.model.carmovevalue.randomnumbergenerator;
 
 public interface NumberGenerator {
     int generateNumber(final int boundaryNumber);
