--- src/main/java/racingcargame/model/movementgenerator/Move.java
@@ -1,5 +1,0 @@
-package racingcargame.model.movementgenerator;
-
-public interface Move {
-    int getMoveValue();
-}
